3-in-1 commit that must be fixed one part fixes a build error in cpufreq. one part belongs to a lib/ backport from 3.10 all the rest is power-efficient wqs implementation.

+++ drivers/cpufreq/cpufreq_governor.c
@@ -356,7 +356,7 @@ int cpufreq_governor_dbs(struct cpufreq_policy *policy,
 					       &j_cdbs->prev_cpu_wall, io_busy);
 			j_cdbs->prev_load = 100 * (unsigned long)(j_cdbs->prev_cpu_wall -
 						   j_cdbs->prev_cpu_idle) /
-						   j_cdbs->prev_cpu_wall;
+						   (unsigned long)j_cdbs->prev_cpu_wall;
 
 			if (ignore_nice)
 				j_cdbs->prev_cpu_nice =
